+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Changed
 - flow event listener for NC28
+- css adjustment for NC28
+- scss removal (<26)
 
 ## 4.11.1 - 2023-11-20
 ### Fixed