+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
+## 4.11.1 - 2023-11-20
 ### Changed
 - flow event listener for NC28
 - css adjustment for NC28
 - scss removal (<26)
+
+### Fixed
+- missing back button when no data sets exist yet
 
 ## 4.11.1 - 2023-11-20
 ### Fixed